+++ tools/gh-actions/resolve-demo-params/resolve-ref.js
@@ -1,0 +1,20 @@
+module.exports = async ({ context, core, inputs }) => {
+    const semver = require('semver');
+
+    const versionInput = inputs['version'];
+
+    if (versionInput === '') {
+        core.info('Version input is empty');
+        return context.ref;
+    } else if (semver.valid(versionInput)) {
+        core.info('Resolving as @alfalab/core-components version');
+
+        return semver.satisfies(versionInput, '>= 49', { includePrerelease: true })
+            ? `@alfalab/core-components@${semver.clean(versionInput)}`
+            : `v${semver.clean(versionInput)}`;
+    }
+
+    core.info('Using version input as ref');
+
+    return versionInput;
+};